node/requirements.txt
bidict==0.21.4
blinker==1.5.0
certifi==2023.11.17
charset-normalizer==3.0.1
click==8.0.4
dnspython==2.2.1
eventlet==0.33.3
greenlet==2.0.2
h11==0.13.0
idna==3.6
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.27.1
simple-websocket==1.0.0
six==1.16.0
urllib3==1.26.18
websockets==9.1
Werkzeug==2.0.3
wsproto==1.0.0
websocket-client
uvicorn
FastApi
